DATA_MINING_PROJECT/requirements.txt
# Core
numpy>=1.23
pandas>=1.5
scipy>=1.9

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Machine Learning
scikit-learn>=1.2
imbalanced-learn>=0.10

# Data Mining
mlxtend>=0.22

# Semi-supervised / Utilities
pyyaml>=6.0
joblib>=1.2

# Notebooks
jupyter>=1.0
ipykernel>=6.25

# Demo / GUI
streamlit>=1.28
